Improve Regexes for identifying docblock target
If a method or property had an argument, and that argument was
destructured, in quotes, or had special characters inside of quotes, it
would have failed. The identifying regexes will still fail if the
argument's default argument is quoted content which contains a closing parenthesis.

A potential issue with failures in parsing the file was found if the file
contains quoted content which contains /* or */ or ){ and other
similarly important markings.  The file may still parse correctly, but
there is a chance there will be issues. you can replace those characters with the unicode
versions to resolve that issue for now.